SafeguardSessions: Adding a SafeguardSessions folder for currently  a SPS only Power BI connector
In order to be able to move forward with the SPS Powerb BI connector
tasks for Cadence 2023.03, I created a SafeguardSession folder where
the SPS connector-related files would be placed for the time being.
They are placed in a separate folder and not in the root of the
repository because this makes it easier to organize if the SPP team
also starts developing their connector and, after subsequent discussions,
the SPP and SPS teams agree on the deeper details of a common connector.

The connector is currently able to display SPS REST API info endpoint
data in a table after it has been given the SPS url and basic
authentication has been performed.

In order to keep the central connector file as clean as possible, I have
created two module files: one with request generation and one with basic
authentication-related helper functions.

References: azure #365895