Test: Add a test folder for the unit test framework and unit tests.
For unit testing, we use the Microsoft-endorsed Fact unit testing
framework. This has the disadvantage that it is not currently built
into the Power Query SDK, so to make it work, you have to add hundreds
of lines to the end of the test files. This can be circumvented by
creating a new connector in a test folder. The connector is modified as
follows:
	* Delete the contents of the pq file. (You may want to leave
	  the version number.)
	* Delete pictures.
	* Delete the previously deleted items from.proj (this is
	  critical if you're using MSBuild).
	* Import the Fact framework into a.pq file and mark the important
	  functions as "shared".
	* Build it with MSBuild or MakePQX
	* (Modify the unit test framework version number in the .pq file
	  if you want to extend or modify it.)

The advantage of this solution is that the code of a test framework is
stored in one place, so if you want to group unit tests, you don't need
to copy them to the end of each file and just modify them in one place.

The disadvantage is that tests cannot be run from the development
environment directly but only from the command line because PQTest.exe
needs two extension parameters (one is Connector.mez and the other one
is UnitTestFramework.mez) and can only be run from the development
environment with one extension.