SafeguardSessions\Request.pqm: Add logging to be able to log the corresponding details of requests
The request generation has been extended to log the corresponding
request-response data. The logging function is evaluated only when
it is called, so I have wrapped the more important variables in the
logger function, which will definitely be used later.

The logger function needs to be given a text value, which will then
be included in the log file. Here I have used the ValueToText function
from Microsoft's TripPin 8 Diagnostics tutorial, which can convert many
types of data into text to be displayed in the log file.

References: azure #365922