sps: Add option to hide successful test cases
Scope: SafeguardSessions/test/UnitTestFramework.pq

As the conclusion of our discussion about the  output of PQTest, we
decided to reduce it's verbosity and only show failed results by
default. This behaviour can be suppressed by passing an environment
variable `short=false` using the `--environmentConfiguration` option, or
setting a `SHORT=false` variable, when running `make unit-tests` or `make
unit-test` targets.

Actually the value of the variable can be any thing, but "true", due to
it being passed as `Text`.

References: azure #399285