sps: Return request url in REST response metadata
Scope: SafeguardSessions/modules/Request.pqm

In order to be able to return the exact query url in case of an error
during the data fetch process, a more futureproof solution must used
that ties the request url to the response value via metadata so that
the information about which request url does the response belong.

This way, the FetchInfo table's url column can be filled appropriately.

Unfortunately, an error object when raised, cannot be extended with
additional metadata, as a result, the request url must be saved in the
error's detail record.

CustomErrors.RaiseErrorOrReturn optional details parameter is removed
since it is not used in production code. This method always receives a
a response object, in which case it is best to save the whole response
value in the error's detail object.

In addition, it would make the production code more complicated as in
BuildSPSResult, the possible non-exsistence of RequestUrl in the detail
must be handled.

CustomErrors.RaiseErrorOrReturn response parameter type is changed from
any to record since it was any only because of the tests.

Since Request.Generate returns the response as record it is better to
treat it as record.

References: azure #351816