sps: Reorganize the structure of test files
Scope: SafeguardSessions/test/

Tests related to files under the modules directory have been moved to
the Unit folder.

The tests for the SafequerdSessions.pq methods have been moved to the
Integration folder. The test files here were created by splitting the
tests in the TestSafeguardSessions.query.pq test file. This was necessary
to make it easier to review the tests.

References: azure #351705