sps: Add url builder and query transformer
Scope: SafeguardSessions/modules/request/

A new request module subfolder has been added, the contents of which are
as follows:

	* Request.pqm
	* UrlBuilder.pqm
	* QueryParameterTransformer.pqm

Request.pqm was simply moved from the module folder to here.

The url builder can be used to build urls from base urls, paths, and
queries. When assembling the URL, we check the query and its parameters,
modify them if necessary, or raise an error. There are also helper
functions to retrieve information from a URL or to return a default
query.

The query parameter transformer is where the actual parameter
transformation takes place, for certain parameters. These are
qInput -> q, from -> start, to -> end. These require transformation
because they are mainly from user input and need to be converted to text
format.

References: azure #351705