sps: Add session search configuration options to the connector
Scope: SafeguardSessions/SafeguardSessions.pq

At the configuration options site, two options came up:
	* via main interface of the connector
	* via the navigation table

We ended up going with the first option because Power BI doesn't handle
it well if we return from navigation table to navigation table, and we
would definitely need to do that since we want to return with schema.

The search conditions (From, To, and Filter Field-Name) are optional.
If the user does not fill in these fields, Power BI will return them as
null.

Dates given as input will be converted to ISO 8601 date format text, so
that there is no risk that the REST API's "start" and "end" query
parameters will contain a date format that is not supported.

Filter field-value pairs are concatenated with "AND" in the q query
parameter. In the case where only the field is specified in the filter
field-value pair, the user will get back a pop up with the following
error message "Filter field was given without value", and if only the
value is specified, the given field-value pair is ignored in the q query
parameter.

References: azure #351705