sps: Define flat schema
Scope: SafeguardSessions/modules/Schema.pqm

The defined SchemaMap contains the presentable and searchable fields,
excluding the 'vault' field. The 'vault' field has been excluded because it
is not deterministic: it can be an object or a list of records. When the 'vault'
field is fixed on SPS side, it will be included in the schema.

References: azure #412244