sps: Update connector documentation
Scope: README.md

Since the repository will be made public, a more detailed guide on the
connector regarding usage, errors and troubleshooting is needed.

In order to avoid the README.md growing to large, contribution and
development related guide is placed into a separate file. Having a
separate CONTRIBUTION.md may also beneficial in the future as the it
contains general information about connector development and creating
pull requests, however the README.md may become SPS specific when the
connector is extended with capabilities to connect to SPP as well.

References: azure #351609