sps: Change icons
Scope: SafeguardSessions/assets/icons/

The icons and their names had to be replaced, as it was found that the
current icons and their names did not match the policy used by the
company.

The newly added icons have been moved to a different location from where
they were before, to better follow the file structure used by the
connector.