sps: Update connector name
Scope: SafeguardSessions/resources.resx

Based on the discussion with the tech writers and product, the
connector name should be "One Identity Safeguard Power BI Connector".

In titles that appear under Home > Get data, only "One Identity
Safeguard" is written as it is evident that it is a Power BI connector.

Folders and modules are not renamed, since the connector in its current
state is SPS specific. When the connector is extended with capabilities
to fetch data from SPP, folders and modules will be restructured and
renamed eventually.

References: azure #351609