sps: Fix fetch count measure in template
Scope: SafeguardSessions/assets/template/OneIdentitySafeguard.pbit

In case the Sessions table is empty, `fetched_sessions_count` measure
contains the value `Blank` (that is equivalent to null) and this was
displayed in the card view as well.

This is fixed by updating the DAX to the following expression so that
it displays 0 in this case as well.

```
fetched_sessions_count = IF(
    OR(
	CALCULATE(FIRSTNONBLANK(Info[Status], TRUE()), ALL(Info)) = "Error",
	COUNTROWS(RELATEDTABLE(Sessions)) = BLANK()
    ),
    0,
    COUNTROWS(RELATEDTABLE(Sessions))
)
```

Original expression was:

```
fetched_sessions_count = IF(
    CALCULATE(FIRSTNONBLANK(Info[Status], TRUE()), ALL(Info)) = "Error",
    0,
    COUNTROWS(RELATEDTABLE(Sessions))
)
```