sps: Extend ApplySchema with error handling
Scope: SafeguardSessions/modules/schema/SchemaUtils.pqm

Error handling has been introduced in the SchemaUtils.ApplySchema method.
This was necessary because if an error occurs here, only the error itself
is logged, not the SPS response data that caused it, but that's an
important element in finding out what caused the error.

During error handling, the unexpected error is logged, and then a new
manually handled error is raised called SchemaApplyError, which contains
the response that caused the error. The new error generated here is then
handled in Utils.BuildSPSResult, where the response is logged.

References: azure #419489