sps: Implement SPS version check
Scope: SafeguardSessions/SafeguardSessions.pqm

Fetching data from unsupoprted versions of SPS is not supported by
default. By introducing the skipVersionCheck input parameter, we can
instruct the connector to bypass the version check process.

As Power Query M is a functional language, all paths need to have a
return value. For this reason, ValidateSPSVersion utilizes the sessionId
if the version check is successful and data can be fetched.

Using FetchData is not possible, as the value of input variables are evaluated at the
time of calling a function, and introducing a proper callback function
is not worth the time cost.

References: azure #351742