sps: Fix importing large number of sessions
Scope: SafeguardSessions/modules/Search.pqm

Since `Table.Combine` is evaluated eagerly, the connector instead of
returning rows on the go one-by-one, it tried to load all of the
session data into memory which would then be persisted into Power BI's
own database. As result, if there were so many sessions (~1.5 million)
that the data import could not even complete as Power BI Desktop
consumed all of the memory and Windows logged out the user of its
session eventually.

To fix this, `Table.Combine` is replaced with `Table.ExpandTableColumn`
that is evaluated lazily and returns rows on the go one-by-one as
expected, so Power BI Desktop does not over-consume memory.

References: azure #428439