sps: Refine version check
Scope: SafeguardSessions/modules/version/Version.pqm

The previously implemented version check using Text.StartsWith accepted
future relases in some cases, e.g. 7.30 if 7.3 was in the supported
versions list.

A possible solution was to explicitly write 7.3. in the supported
versions list or append a "." in the version check function. Instead of
using this method, a more sophisticated solution has been implemented.

The versions to be compared are split along the "." character, the
common length is defined, and the first common-length number of
elements are compared.