sps: Add error codes
Scope: SafeguardSessions\modules\error

The introduction of error codes is necessary to easily identify errors
regardless of the language of the Power BI interface. When creating
error codes, the following convention has been introduced:

 * HTTP request error codes have been extended with a 0
 * Errors originating from the Power BI Connector start at 10000,
   and there are currently no subcategories for them. (The unexpected
   error is given the largest value in the domain.)

The error codes are given in a separate file so that you can easily
see which error-error code pairs are in each category and which are
already occupied error codes within a category.

No subcategories have been introduced for connector-related errors for
the time being, as they are currently quite diverse and difficult to
classify into specific subcategories.

I transformed the unexpected error with the custom function
'TransformUnexpectedError' in the newly made 'UnexpectedErrorWarper'
module. This function will convert the unexpected error to an 'ErrorBase'
type error because this will help handle all errors in a uniform way.
To do this, I moved the Detail field of the original record to an
internal record field. This was necessary because the value of the
Detail field is not necessarily a record.

References: azure #415680