sps: Fix GetDataFromResponse function meta and error handling
Scope: SafeguardSessions\modules\request\ResponseHandler.pqm

Previously, when the GetDataFromResponse's inner GetRecordSafe function
was recursively called, the meta from the original response input was
not passed on, so when an internal field of the path was not found, the
function could not pass the meta field of the original response
RequestURL to the RequestErrors.NotParsableResponse.

This has been fixed, so the output of GetRecordSafe will contain the
meta of the original respond. For the raised error, we will use the
response value given to GetDataFromResponse because if we use the
input record of GetRecordSafe, we would lose more and more of the
original response during recursion because we only pass on the value
of the internal field that was found.

References: azure #426661