repository: Modify and add new entry to git ignore
Scope: .gitingore

During the introduction of the new project structure, the project files
for each new and old Power Query SDK project will be changed so that
MSBuild generates the output not in an "obj" and a "bin" folder but in
a single "Deploy" folder.

The git ignore file has been updated accordingly. But I also kept the
previous Power Query SDK entries, so that if somebody, for some reason,
kept the default settings for a new Power Query SDK project in the
repository, they wouldn't be added by accident.

For new and old entries, I've set it to not matter whether they start
with a small or capital letter.

References: azure #426090