sps: Relocate UnitTestFramework
Scope: UnitTestFrameWork\

The UnitTestFrameWork.pq for the SPS connector has been moved to a
new UnitTestFrameWork folder in the repository root folder, so that
it can be easily accessed and used by the future SPP connector. This is
useful because UnitTestFrameWork.pq can be used for testing any
connector.

References: azure #426090