sps: Factor out common code
Scope: SafeguardCommon\

A SafeguardCommon code folder has been created in the root folder of the
repository, where code snippets from the SPS connector that are generic
enough to be used in the future SPP connector.

The SafeguardCommon folder follows the new structure of the
SafeguardSessions folder.

In addition, a proj and a pq file (the pq file is in the Src folder)
were added to this folder, to test the functionality of the code in the
Src folder.

The tests in the SPS connector for the common codes have been moved to
the SafeguardCommon/Test/Unit folder.

A project file has also been added to the Test folder to create the
UnitTestFrameWork.mez file for SafeguardCommon. This allows the common
code to be tested without any of the main connectors (SPS or the
future SPP).

The originally planned Request.Generate and
ResponseHandler.ValidateByStatusCode functions were not included
in the common code, because if we outsource ValidateByStatusCode it
would have to use the CommonRequestErrors.ErrorMap, which is used in
the SPS connector to create a new ErrorMap with the HTTP code for
SnapshotQuotaError. So the ValidateByStatusCode function in the common
code would not have been able to handle this. And since Request.Generate
uses ValidateByStatusCode, this should not be included in the common
code too.

References: azure #426090