sps: Add separate troubleshooting guide
Scope: SafeguardSessions/TROUBLESHOOTING.md

By replacing `Table.Combine` with `Table.ExpandTableColumn`, a corner
case is introduced in error handling:

When session data is already imported to Power BI and if during a data
refresh, error occurs in the second or subsequent data fetch request,
the current error handling changes so that the error is not persisted
in the `Sessions` table. Instead, Power BI throws a popup that the data
including only the error message in the popup:

```
Failed to save modifications to the server.
Error returned: 'OLE DB or ODBC error: [Server error]
Error 5000: The source IP responded with a server error."
```

In this case, the user cannot drill down in the Sessions to see the
exact error that was returned by SPS to find out the root cause. For
this reason we list each error we expect in the troubleshooting guide
with a suggestion, so we don't leave the users helpless.

Instead of just creating a separate ERRORS.md or something like that,
the whole "Common errors" and "Troubleshooting" section from the
README.md was factored out since it makes more sense to have these in
one place.

References: azure #428439