sps: Fix not-working links
Scope: SafeguardSessions/README.md

Since there were two links for `License` and `Contribution` (one that
referenced the section within the README and one that referenced the
file), only the section reference worked.

For this reason `License` and `Contribution` sections are renamed so
that inside the document, we can keep the short linking with "[]" to a
section instead of using "[](section-name)" reference.