sps: Update .gitignore
Scope: .gitignore

The future bootstrap script for SafeguardSessions will put the files
it downloads and generates into a .venv folder, but the files in this
folder should not be included in the repository, so it is necessary
to add the .venv folder to .gitignor.

The makefile has a target checksum that calculates the sha256 hash
value of the template file for SafeguardSessions, but this hash value
should not be included in the repository either, since it should be
updated every time the template changes. So an entry has been added
to .gitignore that files ending with .sha256 will not be included in
the repository.

References: azure #351709