sps: Fix adv. search responses to match reality
Scope: SafeguardSessions/Test/Integration

Even though without these changes the tests would still pass, they are
commited in order to match the way advanced search responses are
returned in reality. This way later we when code changes so that the
tests would only pass if they are setup to match the reality, we won't
need to debug the test failures.

References: azure #428439