sps: Implement TestConnection function
Scope: SafeguardSessions/Src/Modules/SafeguardSessions.pq

To support Power BI service, the connector must implement a TestConnection
function. The TestConnection function ensures that the SPS is reachable via
the installed gateway. However, the gateway currently only supports text and
number values so skipVersionCheck had to be modified.

Another caveat is that only those optional parameters are passed to the
dataSourcePath that are consecutive to the last required parameter.

References: azure #432548